+++ R/activation.R
@@ -7,6 +7,7 @@
 #' @param peak_time Numeric, peak time of the activation function.
 #' @param curvature Numeric, curvature of the activation function.
 #' @param uncertainty Numeric, indicates how noise is introduced in the system.
+#' @param bw_noise Numeric, amount of between-trial noise.
 #' @param diffusion_coef Numeric, diffusion coefficient.
 #' @param time_step Numeric, time step used to numerical approximation.
 #'
@@ -35,7 +36,7 @@ activation <- function (
         time = 0,
         amplitude = 1.5, peak_time = 0, curvature = 0.4,
         uncertainty = c("par_specific", "overall", "brownian"),
-        diffusion_coef = 0.001, time_step = 0.001
+        bw_noise = 0.01, diffusion_coef = 0.001, time_step = 0.001
         ) {
 
     # uncertainty should be one of above
@@ -45,9 +46,9 @@ activation <- function (
 
         # adding some variability in the other parameters
         # variability is currently fixed but could also be estimated
-        amplitude_sim <- stats::rnorm(n = 1, mean = amplitude, sd = 0.01)
-        peak_time_sim <- stats::rnorm(n = 1, mean = peak_time, sd = 0.01)
-        curvature_sim <- stats::rnorm(n = 1, mean = curvature, sd = 0.01)
+        amplitude_sim <- stats::rnorm(n = 1, mean = amplitude, sd = bw_noise)
+        peak_time_sim <- stats::rnorm(n = 1, mean = peak_time, sd = bw_noise)
+        curvature_sim <- stats::rnorm(n = 1, mean = curvature, sd = bw_noise)
 
         activ_inhib <- amplitude_sim *
             exp(-(log(time) - peak_time_sim)^2 / (2 * curvature_sim^2) )
@@ -56,7 +57,7 @@ activation <- function (
 
         activ_inhib <- pmax(amplitude *
             exp(-(log(time) - peak_time)^2 / (2 * curvature^2) ) +
-            stats::rnorm(n = 1, mean = 0, sd = 0.01), 0)
+            stats::rnorm(n = 1, mean = 0, sd = bw_noise), 0)
 
     } else if (uncertainty == "brownian") {