+++ NEWS.md
@@ -4,9 +4,14 @@
 
 # momimi 0.0.2
 
-* Adding the 3-parameter version of TMM (free parameters are the motor execution threshold, the peak time, and the curvature).
+* Adding the 3-parameter version of TMM (free parameters are the relative motor execution threshold, the peak time, and the curvature).
 
 # momimi 0.0.3
 
 * Adding the possibility to include within-trial (diffusive) noise.
 * Fixing the 3-parameter version of the TMM (TMM3).
+
+# momimi 0.0.4
+
+* Fixing the plotting method of momimi_sim.
+* It is now possible to estimate the amount of between-trial variability (noise) in the TMM3 (which thus now has 4 free parameters).